Amended react-hot-button-component
